lms/requirements.txt
virtualenv==20.13.3
Django==3.1.7
django-storages==1.11.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0